// packages/ui/src/pages/asset-inventory/detail/libs/asset-inventory-detail.types.ts
export type ListMaintainerParams = {
  page: number
  paginate: number
  keyword?: string
  status?: number
}

export type ListAssetElectricityParams = {
  page: number
  paginate: number
  keyword?: string
}

export type ListAssetInventorySupportParams = {
  page: number
  paginate: number
  keyword?: string
}